+++ app/services/save_answer.rb
@@ -17,6 +17,7 @@ class SaveAnswer
 
     if answer.valid?
       answer.save
+      answer.step.check_to_show_additional_step!
       result.success = true
     end