+++ spec/features/visitors/anyone_can_complete_a_journey_spec.rb
@@ -168,6 +168,18 @@ feature "Anyone can start a journey" do
         end
       end
     end
+
+    context "when Contentful entry includes a 'show additional question' rule" do
+      scenario "additional question is shown to the user" do
+        start_journey_from_category_and_go_to_question(category: "show-additional-question.json")
+
+        choose("School expert")
+
+        click_on(I18n.t("generic.button.next"))
+
+        expect(page).to have_content("You should NOT be able to see this question?")
+      end
+    end
   end
 
   context "when the Contentful model is of type staticContent" do